Rename and move PipelineSchemaTableUtil
